Fixes heractic upload cells updates
