+++ piwigo/Settings.bundle/Root.plist
@@ -8,7 +8,7 @@
 			<key>Type</key>
 			<string>PSGroupSpecifier</string>
 			<key>FooterText</key>
-			<string>© 2015-2024 Spencer Baker, Olaf Greck, Eddy Lelièvre-Berna.</string>
+			<string>© 2015-2025 Spencer Baker, Olaf Greck, Eddy Lelièvre-Berna.</string>
 		</dict>
 		<dict>
 			<key>Type</key>
